requirements.txt
asgiref==3.11.0
charset-normalizer==3.4.4
Django
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
# mysqlclient==2.2.7
pillow==12.1.0
psycopg2-binary==2.9.11
PyJWT
# PyMySQL==1.1.2
python-decouple==3.8
qrcode==8.2
reportlab==4.4.9
sqlparse==0.5.5
tzdata==2025.3
drf-yasg
whitenoise